Initial setup and basic functional Progamm
Added dependencies serde and serde_yaml in the Cargo.toml file to enable working with YAML. Created the config folder with a README file explaining the purpose and structure of the folder and its YAML files. The example.yaml and test.yaml files are included for testing purposes.

The main file, main.rs, has been implemented in Rust. It is capable of reading command line arguments and searching the YAML files in the config folder for a match in the "id" field. This marks the completion of the first stage of the program.